unit test and lang added

+++ app/Http/Controllers/Admin/UserController.php
@@ -6,34 +6,48 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use App\Models\User;
 use App\Exceptions\Handlers\UniqueConstraintViolationExceptionHandler;
+use App\Http\Requests\UserRequest;
 use Exception;
 use GuzzleHttp\Psr7\Response;
 use Illuminate\Database\UniqueConstraintViolationException;
 
 class UserController extends Controller
 {
-    public function list_users()
+    public function list_users(Request $request)
     {
+        // dd($request->header());
         $users = User::where('role_id', '=', 2)->paginate(5);
         return Response([
             'success' => true,
             'users' => $users
-        ]);
+        ],200);
     }
 
-    public function update_user(int $user_id, Request $request)
+    public function update_user(int $user_id, UserRequest $request)
     {
         $user = User::whereId($user_id)->first();
+        if(! $user){
+            return Response([
+                'success'=>true,
+                'message'=>"User does not exist.",
+            ],422);
+        }
         $user->update($request->all());
         return Response([
             'success' => true,
-            $user
+            'message' =>'User data updated'
         ], 200);
     }
 
     public function delete_user(int $user_id)
     {
-        $user = User::whereId($user_id);
+        $user = User::whereId($user_id)->first();
+        if(! $user){
+            return Response([
+                'success'=>true,
+                'message'=>"User does not exist.",
+            ],422);
+        }
         $user->delete();
         return Response([
             'success' => true,
@@ -41,24 +55,10 @@ class UserController extends Controller
         ], 200);
     }
 
-    public function add_user(Request $request)
+    public function add_user(UserRequest $request)
     {
         try {
-            $data = $request->validate([
-                'name' => 'required|max:255',
-                'email' => 'required|email:rfc,dns',
-                'age' => 'required|integer',
-                'role_id' => 'required|integer',
-                'password' => 'required|string|max:50',
-                'phone_number' => 'required|string',
-            ]);
-        } catch (Exception $e) {
-            return Response([
-                'success' => false,
-                'message' => "Invalid argument"
-            ], 400);
-        }
-        try {
+            // dd($request->only(['role_id']));SSS
             $newUser = User::create([
                 'name' => $request->input('name'),
                 'email' => $request->input('email'),
@@ -71,7 +71,7 @@ class UserController extends Controller
             return Response([
                 'success' => true,
                 'users' => $newUser
-            ], 200);
+            ], 201);
         } catch (UniqueConstraintViolationException $exception) {
             $handler = app()->make(UniqueConstraintViolationExceptionHandler::class);
             return $handler->handle($exception);
